refactor(issues): eslint,svg,prettier
Resolves: #1, #2, #3
