+++ src/App.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { ref } from 'vue';
-import Header from '@/components/Header.vue';
+import Header from '@/components/HeaderComponent.vue';
 
 const msg = ref('PizzaOrder');
 </script>
